requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
playwright==1.41.1
xlsxwriter==3.1.9
jinja2==3.1.3
aiofiles==23.2.1
Pillow
email-validator